+++ tests/test_enums.py
@@ -1,6 +1,7 @@
 import json
 import unittest
 
+from inttegro.otp import Purpose as OTPPurpose
 from inttegro.product import Type as ProductType
 from inttegro.refund import Reason as RefundReason
 from inttegro.upload_request import Status as UploadRequestStatus
@@ -11,13 +12,14 @@ class ApiEnumTest(unittest.TestCase):
         encoded = json.dumps(
             {
                 "product": ProductType.DIGITAL,
+                "otp": OTPPurpose.SIGN_IN,
                 "refund": RefundReason.REQUESTED_BY_CUSTOMER,
                 "status": UploadRequestStatus.PENDING,
             }
         )
         self.assertEqual(
             encoded,
-            '{"product": "digital", "refund": "requested_by_customer", "status": "pending"}',
+            '{"product": "digital", "otp": "sign_in", "refund": "requested_by_customer", "status": "pending"}',
         )